solution/scraper/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.42
psycopg2-binary==2.9.10
requests==2.31.0
beautifulsoup4==4.12.2
celery==5.3.4
redis==5.0.1
python-multipart==0.0.6
flower==2.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
